+++ packages/web/controls/__stories__/TextInputPerformance.stories.tsx
@@ -9,7 +9,6 @@ export default {
   component: TextInput,
   decorators: [withPerformance],
   excludeStories: [
-    'CDSTextInput',
     'HTMLInput',
     'HundredCDSComponents',
     'HundredHTMLComponent',